1. fixed NtCreateProcess to do some buffer checks and moved the implementation to an internal function so it can be shared with PsCreateSystemProcess(). Also don't leak so many resources on failures 2. processes should acuire a cid handle for their unique process id 3. fixed several instances in structures where process ids were DWORD/ULONG instead of HANDLEs
svn path=/trunk/; revision=13301

+++ reactos/lib/kernel32/process/proc.c
@@ -324,7 +324,7 @@ GetProcessId(HANDLE Process)
     return 0;
   }
 
-  return ProcessBasic.UniqueProcessId;
+  return (DWORD)ProcessBasic.UniqueProcessId;
 }
